+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/impl/BaseClientTest.java
@@ -46,8 +46,9 @@ class BaseClientTest {
   void setUp() {
     httpClient = mock(HttpClient.class, RETURNS_DEEP_STUBS);
     baseClient = new BaseClient(httpClient, Config.empty(), () -> Runnable::run) {
+
       @Override
-      BaseClient newInstance(Config config) {
+      BaseClient copy() {
         return baseClient;
       }